+++ front/src/pages/board/ListPage.jsx
@@ -1,70 +1,12 @@
 import React from 'react';
-import { useLocation } from 'react-router-dom'; // 현재 경로 가져오기
-// import { Outlet } from "react-router-dom";
-import BasicLayout from "../../layouts/BasicLayout";
-import Sidebar from "../../components/common/Sidebar";
-import ListTitle from "../../components/board/list/ListTitle";
-
 
 const ListPage = () => {
 
-    const location = useLocation();
-
-    // console.log(location);
-    
-    let sub= "";
-    let info= "";
-
-    switch (location.pathname) {
-        case '/board/':
-            sub = "전체";
-            info="자취 고수들의 공유마당을 둘러보고 꿀팁을 얻어가세요!"
-            break;
-        case '/board/restaurants/':
-            sub = "맛집";
-            info="내 자취방 근처에 맛집이?"
-            break;
-        case '/board/clean/':
-            sub = "청소";
-            info="오늘도 열심히 깨끗하게! 자신있게!"
-            break;
-        case '/board/recipe/':
-            sub = "요리";
-            info="오늘은 내가 바로 셰프! 멋진 실력을 보여주세요!"
-            break;
-        case '/board/wealth/':
-            sub = "재테크";
-            info="텅장만 갖고 있지말고 불려보세요!"
-            break;
-        case '/board/interior/':
-            sub = "인테리어";
-            info="이왕이면 예쁘게 인테리어하고 싶다면?"
-            break;
-        case '/board/policy/':
-            sub = "정책";
-            info="나만 몰라, 청년정책! 도와줘요~"
-            break;
-        case '/board/etc/':
-            sub = "기타";
-            info="더 자세하고 다양한 꿀팁 집합소!"
-            break;
-        default:
-            sub = "전체";
-            info="자취 고수들의 공유마당을 둘러보고 꿀팁을 얻어가세요!"
-            break;
-    }
-
     return (
-        <BasicLayout>
-            <Sidebar pageType="board" />
-            <div className="contents">
-                <ListTitle main="꿀팁" sub={sub} info={info}/>
-                콘텐츠 부분
-                {/* <Outlet /> */}
-            </div>
-        </BasicLayout>
+        <div className="contents">
+            콘텐츠 부분
+        </div>
     );
 }
-
 
 export default ListPage;
