+++ front/src/router/boardRouter.jsx
@@ -1,4 +1,5 @@
 import { Suspense, lazy } from "react";
+import { Navigate } from "react-router-dom";
 
 const Loading = <div>Loading....</div>
 
@@ -6,6 +7,14 @@ const ListPage = lazy(() => import("../pages/board/ListPage"))
 
 const boardRouter = ()=> {
     return [
+        {
+            path: "",
+            element: <Navigate replace to="/board/total"/>
+        },
+        {
+            path: "total",
+            element: <Suspense fallback={Loading}><ListPage/></Suspense>
+        },
         {
             path: "restaurants",
             element: <Suspense fallback={Loading}><ListPage/></Suspense>
