+++ front/src/router/matchingRouter.jsx
@@ -1,4 +1,5 @@
 import { Suspense, lazy } from "react";
+import { Navigate } from "react-router-dom";
 
 const Loading = <div>Loading....</div>
 
@@ -6,6 +7,14 @@ const ListPage = lazy(() => import("../pages/matching/ListPage"))
 
 const matchingRouter = ()=> {
     return [
+        {
+            path: "",
+            element: <Navigate replace to="/matching/total"/>
+        },
+        {
+            path: "total",
+            element: <Suspense fallback={Loading}><ListPage/></Suspense>
+        },
         {
             path: "regularly",
             element: <Suspense fallback={Loading}><ListPage/></Suspense>
